% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_by_category.R
\name{count_by_category}
\alias{count_by_category}
\title{**count_by_category**}
\usage{
count_by_category(data, column_name, limit, is_descending = TRUE)
}
\arguments{
\item{data:}{input data frame that you want to select from. Here naming of this parameter is self-explanatory.}

\item{column_name:}{the column we want to categorize against. This parameter could be named as column as well.}

\item{limit:}{the number of records that we want to return. The parameter name limit is self-documenting and clearly conveys API specification.}

\item{is_descending:}{Optional boolean parameter which is set to TRUE by default.}
}
\value{
a table, counting the number of observations per category.
}
\description{
Given a data frame, for a specific column, count of the number of observations per category. This function removes NA values
and then count per category. By default the function return result in descending order.
}
\examples{

count_by_category(apt_buildings, property_type, 5, )
count_by_category(apt_buildings, heating_type, 5, TRUE)
count_by_category(building_permits, property_use, 5, TRUE)
count_by_category(building_permits, property_use, 5, TRUE)
count_by_category(steam_games, developer, 6)
count_by_category(steam_games, developer, 5, FALSE)
}
