% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_by_category.R
\name{is.boolean}
\alias{is.boolean}
\title{is.boolean function is motivated from https://github.com/Rapporter/}
\usage{
is.boolean(x)
}
\arguments{
\item{x}{an object to check}
}
\value{
a logical value indicating whether provided object is a boolean
}
\description{
Checks if provided object is a boolean i.e. a length-one logical vector.
}
